+++ gameserver/logic/skills/TransformOnDeath.cs
@@ -12,7 +12,7 @@ namespace LoESoft.GameServer.logic.behaviors
         private readonly int max;
         private readonly int min;
         private readonly float probability;
-        private readonly ushort target;
+        private readonly int target;
         private readonly bool returnToSpawn;
 
         public TransformOnDeath(
